Fixed NumberOfEvents filter.
If applied, this commit will have fixed the NumberOfEvents filtered according the the input.
For this was added a new state "currentLocation" that accepts the user "location" in updateEvents method.
Using slice() was added the state of "numberOfEvents" to "events" state that contains the "event.location".
Was created another method "updateNumberOfEvents" that set the state of the input(eventCount) value and update the "updateEvents" state adding the "currentLocation" and "numberOfEvents" own state (inputed number).
Afterwards this methods have been passed as props in NumberOfEvents.